依赖升级

+++ eva-web/eva-web-sys/src/main/java/io/nerv/web/sys/dict/service/DictService.java
@@ -163,7 +163,7 @@ public class DictService extends StdBaseService<DictMapper, DictEntity> {
      * @return
      */
     public boolean checkUnique(DictEntity dictEntity) {
-        int records = this.mapper.selectCount(new QueryWrapper<>(dictEntity));
-        return records>0;
+        long records = this.mapper.selectCount(new QueryWrapper<>(dictEntity));
+        return records > 0;
     }
 }
